Multi Signature wallets
Multisignature wallets are cryptocurrency wallets that require two or more private keys to sign and send a transaction, This is a digital wallet that operates with multi-signature addresses. requires more than one private key to sign and authorize a crypto transaction 
or, in some cases, several different keys can be used to generate a signature.