A simple Escrow Account implement with Solidity
A crypto escrow service implemented in solidity solves this conundrum by safely keeping your crypto in the deposit account to protect both the buyer and the seller. The crypto is placed in the control of an independent third party. This helps keep the transactions more secure by keeping the payment safe until all the terms have been met.